button protocol and driver: Add a new protocol and driver for a Button
A button can be used to simulate pressing a physical button on a device. For
example a DigitalOutputButtonDriver can be used when a reset button is
connected via a relay to a DigitalOutput.  In addition there are the
ManualButtonDriver and ExternalButtonDriver drivers analogous to the
ManualPowerDriver and ExternalPowerDriver drivers.

Operations on a button include "press", "release", "press_for", and "get".

Signed-off-by: Perry Melange <isprotejesvalkata@gmail.com>