LibGPIO: Add the LibGPIO resources and update the SysfsGPIO resources
Add the LibGPIO, NetworkLibGPIO and MatchedLibGPIO resources.  These
resouces bind to the /dev/gpiochipN and it's associated gpio line.  In
addition there is an optional active_low attribute which can be used to
invert the logical value used on the gpio line.

The resources SysfsGPIO, NetworkSysfsGPIO and MatchedSysfsGPIO resources
have been modified to have an additional optional active_low attribute
which can be used to invert the logical value used on the gpio line.

The SysfsGPIO, NetworkSysfsGPIO and MatchedSysfsGPIO resouces are now marked
as deprecated.  See https://www.kernel.org/doc/Documentation/gpio/sysfs.txt

Signed-off-by: Perry Melange <isprotejesvalkata@gmail.com>